poc-django-celery/requirements.txt
celery[redis]==4.2.0
flower==0.9.3
Django==2.0.7
djangorestframework==3.8.2
requests==2.19.1
